v0.6
Bug with not updating training group details was fixed. Also react-firebase-hooks were implemented. The chat architecture was changed for optimised perfomance. I also have added triggers for updating the avatar for the user in chat and training-groups. Personal chats and group chats are working ok (on top of that I added the function to delete a personal chat)